Change environ vars file to match virtualenv spec
When starting virtualenv, virtualenv looks for a file called
'postactivate' in {{ virtualenv_dir }}/bin. Putting environment
variables fits in with virtualenv's way of doing things. This makes it
easier in development enivironments because all a person has to do is
type workon {{ virtualenv_name }} and virtualenv will automatically
source the postactivate script.